// lib/types.ts
import { type Message } from 'ai'

export interface Chat extends Record<string, any> {
  id: string
  title: string
  createdAt: Date
  userId: string
  path: string
  messages: Message[]
  sharePath?: string
  conversationSoFar: string,
  previouslySuggestedRestaurants: string[],
  groupId: string,
  groupChatDetailsId: string,
}

export type ServerActionResult<Result> = Promise<
  | Result
  | {
      error: string
    }
>